configs for final baseline runs

+++ models/t5_utils.py
@@ -160,7 +160,9 @@ def print_trainable_parameters(model):
 def t5_init(model_checkpoint, tokenizer_checkpoint):
     """initialize model and tokenizer, mode=Default, LoRA"""
 
-    model = T5ForConditionalGeneration.from_pretrained(model_checkpoint)
+    model = T5ForConditionalGeneration.from_pretrained(
+        model_checkpoint, torch_dtype=torch.bfloat16
+    )
     tokenizer = T5Tokenizer.from_pretrained(tokenizer_checkpoint)
 
     # if lora:
